+++ client/src/containers/AddContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import Button from '../styled-components/Button';
-import AddArtist from './AddArtist';
-import AddGroup from './AddGroup';
+import AddArtist from '../components/AddArtist';
+import AddGroup from '../components/AddGroup';
 import { Group, Person } from '../types/ArtistTypes';
 
 interface Props {
